+++ customtkinter/windows/widgets/__init__.py
@@ -1,3 +1,19 @@
-from customtkinter.windows.widgets.core_rendering.ctk_canvas import CTkCanvas
+from .ctk_button import CTkButton
+from .ctk_checkbox import CTkCheckBox
+from .ctk_combobox import CTkComboBox
+from .ctk_entry import CTkEntry
+from .ctk_frame import CTkFrame
+from .ctk_label import CTkLabel
+from .ctk_optionmenu import CTkOptionMenu
+from .ctk_progressbar import CTkProgressBar
+from .ctk_radiobutton import CTkRadioButton
+from .ctk_scrollbar import CTkScrollbar
+from .ctk_segmented_button import CTkSegmentedButton
+from .ctk_slider import CTkSlider
+from .ctk_switch import CTkSwitch
+from .ctk_tabview import CTkTabview
+from .ctk_textbox import CTkTextbox
 
+# init canvas font character map for current platform
+from .core_rendering import CTkCanvas
 CTkCanvas.init_font_character_mapping()
