+++ customtkinter/windows/widgets/ctk_button.py
@@ -2,12 +2,12 @@ import tkinter
 import sys
 from typing import Union, Tuple, Callable, Optional
 
-from .core_rendering.ctk_canvas import CTkCanvas
-from .theme.theme_manager import ThemeManager
-from .core_rendering.draw_engine import DrawEngine
-from .core_widget_classes.widget_base_class import CTkBaseClass
-from .font.ctk_font import CTkFont
-from .image.ctk_image import CTkImage
+from .core_rendering import CTkCanvas
+from .theme import ThemeManager
+from .core_rendering import DrawEngine
+from .core_widget_classes import CTkBaseClass
+from .font import CTkFont
+from .image import CTkImage
 
 
 class CTkButton(CTkBaseClass):
