+++ customtkinter/windows/widgets/ctk_progressbar.py
@@ -2,10 +2,10 @@ import tkinter
 import math
 from typing import Union, Tuple, Optional
 
-from .core_rendering.ctk_canvas import CTkCanvas
-from .theme.theme_manager import ThemeManager
-from .core_rendering.draw_engine import DrawEngine
-from .core_widget_classes.widget_base_class import CTkBaseClass
+from .core_rendering import CTkCanvas
+from .theme import ThemeManager
+from .core_rendering import DrawEngine
+from .core_widget_classes import CTkBaseClass
 
 
 class CTkProgressBar(CTkBaseClass):
