+++ customtkinter/windows/widgets/ctk_radiobutton.py
@@ -2,11 +2,11 @@ import tkinter
 import sys
 from typing import Union, Tuple, Callable, Optional
 
-from .core_rendering.ctk_canvas import CTkCanvas
-from .theme.theme_manager import ThemeManager
-from .core_rendering.draw_engine import DrawEngine
-from .core_widget_classes.widget_base_class import CTkBaseClass
-from .font.ctk_font import CTkFont
+from .core_rendering import CTkCanvas
+from .theme import ThemeManager
+from .core_rendering import DrawEngine
+from .core_widget_classes import CTkBaseClass
+from .font import CTkFont
 
 
 class CTkRadioButton(CTkBaseClass):
