+++ customtkinter/windows/widgets/utility/__init__.py
@@ -1,0 +1,1 @@
+from .utility_functions import pop_from_dict_by_set, check_kwargs_empty
